[#15] Moved components to containers folder. Now to split up code.
